requirements.txt
# PaddleOCR 异步识别桌面工具 核心依赖
# 基于用户环境的 'pip freeze' 输出，保留了 GPU 支持和核心功能依赖。

# ----------------------------------------------------
# 1. 核心 OCR 引擎 (包含 GPU 版本)
# ----------------------------------------------------
# 使用 GPU 版本以确保环境重现
paddlepaddle-gpu==3.2.0
paddleocr==3.2.0

# ----------------------------------------------------
# 2. GUI、图像处理和配置
# ----------------------------------------------------
# Pillow (PIL) - 用于图像处理和 Tkinter 预览/图标
Pillow==11.3.0
# PyYAML - 用于读取 config.yaml 配置文件
PyYAML==6.0.2

# ----------------------------------------------------
# 3. 截图功能依赖
# ----------------------------------------------------
# mss 和 pyscreenshot 是常用的截图库，确保截图工具运行
mss==10.1.0
pyscreenshot==3.1

# ----------------------------------------------------
# 4. 必要的运行时依赖 (用于 GPU 和 OCR 运行)
# ----------------------------------------------------
numpy==2.2.6
opencv-contrib-python==4.10.0.84
scipy==1.15.3
six==1.17.0
tqdm==4.67.1

# ----------------------------------------------------
# 5. NVIDIA/CUDA 运行时依赖 (确保 GPU 环境重现)
# ----------------------------------------------------
nvidia-cublas-cu12==12.9.0.13
nvidia-cuda-runtime-cu12==12.9.37
nvidia-cudnn-cu12==9.9.0.52
nvidia-cufft-cu12==11.4.0.6
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.4.40
nvidia-cusparse-cu12==12.5.9.5
nvidia-nvjitlink-cu12==12.9.86

# ----------------------------------------------------
# 6. 其他辅助依赖
# ----------------------------------------------------
requests==2.32.5
protobuf==6.31.1
typing_extensions==4.15.0